requirements.txt
ujson==1.35
cython==0.29.3
gunicorn==19.9.0
pandas==0.23.4
requests==2.21.0
tqdm==4.29.1
pytest==4.1.1
pytest-cov==2.6.1
locustio==0.9.0